+++ src/anyio/_core/_tasks.py
@@ -6,7 +6,7 @@ from collections.abc import (
     Coroutine,
     Generator,
 )
-from contextlib import AbstractContextManager, contextmanager
+from contextlib import contextmanager
 from enum import Enum, auto
 from inspect import iscoroutine
 from types import TracebackType
@@ -198,9 +198,7 @@ def move_on_at(deadline: float | None, shield: bool = False) -> CancelScope:
     )
 
 
-def move_on_after(
-    delay: float | None, shield: bool = False
-) -> AbstractContextManager[CancelScope]:
+def move_on_after(delay: float | None, shield: bool = False) -> CancelScope:
     """
     Create a cancel scope with a deadline that expires after the given delay.
 
@@ -301,11 +299,16 @@ class TaskHandle(Generic[T_co, T_startval_co]):
     _return_value: T_co
     _start_value: T_startval_co
 
-    def __init__(self, coro: Coroutine[Any, Any, T_co], name: object) -> None:
+    def __init__(
+        self,
+        coro: Coroutine[Any, Any, T_co],
+        name: object,
+        cancel_scope: CancelScope | None = None,
+    ) -> None:
         from ._synchronization import Event
 
         self._coro = coro
-        self._cancel_scope = CancelScope()
+        self._cancel_scope = cancel_scope if cancel_scope is not None else CancelScope()
         self._finished_event = Event()
         self._exception: BaseException | None = None