+++ src/anyio/_core/_tempfile.py
@@ -17,6 +17,7 @@ from typing import (
 from .. import to_thread
 from .._core._fileio import AsyncFile
 from ..lowlevel import checkpoint_if_cancelled
+from ._tasks import CancelScope
 
 if TYPE_CHECKING:
     from _typeshed import OpenBinaryMode, OpenTextMode, ReadableBuffer, WriteableBuffer
@@ -505,9 +506,10 @@ class TemporaryDirectory(Generic[AnyStr]):
         traceback: TracebackType | None,
     ) -> None:
         if self._tempdir is not None:
-            await to_thread.run_sync(
-                self._tempdir.__exit__, exc_type, exc_value, traceback
-            )
+            with CancelScope(shield=True):
+                await to_thread.run_sync(
+                    self._tempdir.__exit__, exc_type, exc_value, traceback
+                )
 
     async def cleanup(self) -> None:
         if self._tempdir is not None: